+++ NEWS.md
@@ -6,6 +6,7 @@ across sessions.
 locking variables when using `set`.
 - Better instructions for deploying as a service.
 - Remove the deprecated `Logger` class, see [log](https://github.com/devOpifex/log) package.
+- Pass `host` to free port fetch function.
 
 # ambiorix 1.0.2